+++ deser-json/tests/integration.rs
@@ -5,5 +5,6 @@ mod test_bytes;
 mod test_de;
 mod test_locations;
 mod test_nesting;
+mod test_pretty;
 mod test_ser;
 mod test_stream;